[♻️refactor]: 페이지 내에서만 사용되는 타입은 페이지 내부에 위치
