+++ src/services/kitOrder/postOrderPurchase.ts
@@ -1,5 +1,5 @@
 import { LOCAL_STORAGE_AUTH_TOKEN } from '@/constants/localStorageKey';
-import { PurchaseResponse } from '@/pages/KitOrderPage/type';
+import { PurchaseResponse } from '@/pages/Kit/KitOrderPage/type';
 
 import { API_AUTHORITY, API_ENDPOINT } from '../apiEndpoint';
 import { axiosInstance } from '../axiosInstance';
