--- src/types/KitDetail/kitDetail.ts
@@ -1,5 +1,0 @@
-import { AxiosResponse } from 'axios';
-
-import { IKit } from '@/pages/KitDetailPage/type';
-
-export type KitResponse = AxiosResponse<IKit>;
